bindings: containers/checkpoint uses entities/types
