+++ pkg/bindings/containers/checkpoint.go
@@ -7,13 +7,13 @@ import (
 	"os"
 
 	"github.com/containers/podman/v4/pkg/bindings"
-	"github.com/containers/podman/v4/pkg/domain/entities"
+	"github.com/containers/podman/v4/pkg/domain/entities/types"
 )
 
 // Checkpoint checkpoints the given container (identified by nameOrID).  All additional
 // options are options and allow for more fine grained control of the checkpoint process.
-func Checkpoint(ctx context.Context, nameOrID string, options *CheckpointOptions) (*entities.CheckpointReport, error) {
-	var report entities.CheckpointReport
+func Checkpoint(ctx context.Context, nameOrID string, options *CheckpointOptions) (*types.CheckpointReport, error) {
+	var report types.CheckpointReport
 	if options == nil {
 		options = new(CheckpointOptions)
 	}
@@ -52,13 +52,13 @@ func Checkpoint(ctx context.Context, nameOrID string, options *CheckpointOptions
 		return nil, err
 	}
 
-	return &entities.CheckpointReport{}, nil
+	return &types.CheckpointReport{}, nil
 }
 
 // Restore restores a checkpointed container to running. The container is identified by the nameOrID option. All
 // additional options are optional and allow finer control of the restore process.
-func Restore(ctx context.Context, nameOrID string, options *RestoreOptions) (*entities.RestoreReport, error) {
-	var report entities.RestoreReport
+func Restore(ctx context.Context, nameOrID string, options *RestoreOptions) (*types.RestoreReport, error) {
+	var report types.RestoreReport
 	if options == nil {
 		options = new(RestoreOptions)
 	}
