+++ pkg/domain/entities/types/containers.go
@@ -9,3 +9,19 @@ type ContainerCopyFunc func() error
 type ContainerStatReport struct {
 	define.FileInfo
 }
+
+type CheckpointReport struct {
+	Err             error                                   `json:"-"`
+	Id              string                                  `json:"Id"` //nolint:revive,stylecheck
+	RawInput        string                                  `json:"-"`
+	RuntimeDuration int64                                   `json:"runtime_checkpoint_duration"`
+	CRIUStatistics  *define.CRIUCheckpointRestoreStatistics `json:"criu_statistics"`
+}
+
+type RestoreReport struct {
+	Err             error                                   `json:"-"`
+	Id              string                                  `json:"Id"` //nolint:revive,stylecheck
+	RawInput        string                                  `json:"-"`
+	RuntimeDuration int64                                   `json:"runtime_restore_duration"`
+	CRIUStatistics  *define.CRIUCheckpointRestoreStatistics `json:"criu_statistics"`
+}
